refactor: 💡 Reogranize sub-component files into folders
