+++ App.tsx
@@ -1,8 +1,8 @@
 import { StatusBar } from "expo-status-bar";
 import { ScrollView } from "react-native";
 import { styles } from "./styles/styles";
-import AppHeader from "./components/AppHeader";
-import Calendar from "./components/Calendar";
+import AppHeader from "./components/header/AppHeader";
+import Calendar from "./components/calendar/Calendar";
 
 export default function App() {
   return (
